backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2025.8.3",
    "fastapi>=0.116.1",
    "pymongo>=4.14.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
